moved createRooms to RoomsController and added getting rooms as csv in edit screen
TODO: fetched rooms are recreated so they are doubled in the db, FIX
